reading `TERESA_ENVIRONMENT` and database conf from environment
database vars are: `TERESADB_HOSTNAME`, `TERESADB_PORT`, `TERESADB_USERNAME`,
`TERESADB_PASSWORD`, `TERESADB_DATABASE`

If those variables are set, we try and connect on the mysql listening
on that respective address. If not, we use an sqlite.

If `TERESA_ENVIRONMENT`=`PRODUCTION` we panic if those variables aren't set
OR if we fail to connect to the database.