add current env to app info
It is possible for users to manipulate an APP in the wrong environment
and to avoid that I suggest adding the current environment when listing
the APP info.

A possible downside of this change is the fact that it is using the
function `ReadConfigFile (cfgFile)) (*Config, error)` to get the current
cluster name.